Upon further investigation I found some sloppy coding in BotAction. I decided to make my isBinary method in Tools a bit more fancy and stop !doors from killing the whole util if you enter a wrong value. Huzzah.
git-svn-id: http://twcore.org/svn/trunk@2286 9f8b3893-2a26-0410-9d00-a71e40fd2f77

+++ twcore/src/twcore/core/BotAction.java
@@ -1573,7 +1573,7 @@ public class BotAction
     {
         try
         {
-            setDoors((int) Byte.parseByte(eightBinaryDigits, 2));
+            setDoors((int) Short.parseShort(eightBinaryDigits, 2));
         }
         catch (NumberFormatException e)
         {
